backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
python-dotenv
werkzeug
PyJWT
Flask-SocketIO
eventlet
